refactor: 여러 컨디션을 처리하는 핸들러들의 중복 코드 제거
